fix some merge oddities
